requirements.txt
Flask==2.3.3
asgiref==3.7.2
Flask-CORS==4.0.0
langchain
scikit-learn
langgraph
langchain-groq
python-dotenv==1.0.0
gitpython==3.1.40
requests==2.31.0
pygithub==2.1.1
radon==6.0.1
pylint==3.0.3
black==23.12.1
astroid==3.0.2
chardet==5.2.0
pygments==2.17.2
markdown==3.5.1
plotly==5.18.0
pandas==2.1.4
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0